feat: add data-type attribute to background images for lazy loading
This commit adds a data-type attribute to the background images in the HTML file. This attribute is used for lazy loading the images based on the device type (desktop or mobile). The CSS file is also updated to target the images based on the data-type attribute and control their display.