`	ext feat: optimize lazy loading of background images and add data-type attribute
This commit optimizes the lazy loading of background images by adding a data-type attribute to the HTML file. The data-type attribute is used to control the display of the images based on the device type (desktop or mobile). The CSS file is also updated to target the images based on the data-type attribute.